requirements.txt
pandas
numpy
scipy
flask
google.generativeai
python-dotenv
